requirements.txt
pyspark
delta-spark
duckdb
polars